+++ src/pages/programsWip/form/programStage/StageForm.tsx
@@ -38,7 +38,6 @@ import {
     PickWithFieldFilters,
     ProgramStage,
 } from '../../../../types/models'
-import { ProgramStageListItem } from '../ProgramStagesFormContents'
 import styles from './StageForm.module.css'
 import { StageFormContents } from './StageFormContents'
 import { StageFormDescriptor } from './stageFormDescriptor'
@@ -52,6 +51,7 @@ export const fieldFilters = [
     'style[color,icon]',
     'enableUserAssignment',
     'featureType',
+    'validationStrategy',
     'preGenerateUID',
     'executionDateLabel',
     'dueDateLabel',
@@ -103,15 +103,9 @@ export type StageFormProps = {
     stage?: PartialStageFormValues
     onCancel?: () => void
     onSubmit: OnSubmitWithClose
-    existingStages?: ProgramStageListItem[]
 }
 
-export const StageForm = ({
-    stage,
-    onSubmit,
-    onCancel,
-    existingStages,
-}: StageFormProps) => {
+export const StageForm = ({ stage, onSubmit, onCancel }: StageFormProps) => {
     const programId = useParams().id as string
     const customAttributes = useCustomAttributesQuery({
         enabled: true,
@@ -166,7 +160,6 @@ export const StageForm = ({
                                             setSelectedSection={
                                                 setSelectedSection
                                             }
-                                            existingStages={existingStages}
                                         />
                                         <SectionedFormErrorNotice />
                                     </div>
@@ -226,7 +219,6 @@ export const EditStageForm = ({
     stage,
     onCancel,
     onSubmitted,
-    existingStages,
 }: {
     stage: DisplayableModel
     onCancel: () => void
@@ -234,7 +226,6 @@ export const EditStageForm = ({
         values: SubmittedStageFormValues,
         closeOnSubmit: boolean
     ) => void
-    existingStages?: ProgramStageListItem[]
 }) => {
     const handlePatch = usePatchModel(stage.id, stageSchemaSection.namePlural)
 
@@ -339,7 +330,6 @@ export const EditStageForm = ({
             stage={stageValues.data}
             onSubmit={onFormSubmit}
             onCancel={onCancel}
-            existingStages={existingStages}
         />
     )
 }
@@ -347,14 +337,12 @@ export const EditStageForm = ({
 export const NewStageForm = ({
     onCancel,
     onSubmitted,
-    existingStages,
 }: {
     onCancel: () => void
     onSubmitted: (
         values: SubmittedStageFormValues,
         closeOnSubmit: boolean
     ) => void
-    existingStages?: ProgramStageListItem[]
 }) => {
     const handleCreate = useCreateModel(stageSchemaSection.namePlural)
     const onFormSubmit: OnSubmitWithClose = async (
@@ -384,7 +372,6 @@ export const NewStageForm = ({
             stage={undefined}
             onSubmit={onFormSubmit}
             onCancel={onCancel}
-            existingStages={existingStages}
         />
     )
 }
@@ -393,7 +380,6 @@ export const EditOrNewStageForm = ({
     stage,
     onCancel,
     onSubmitted,
-    existingStages,
 }: {
     stage: DisplayableModel | null | undefined
     onCancel: () => void
@@ -401,20 +387,13 @@ export const EditOrNewStageForm = ({
         values: SubmittedStageFormValues,
         closeOnSubmit: boolean
     ) => void
-    existingStages?: ProgramStageListItem[]
 }) => {
     if (stage === undefined) {
         return null
     }
 
     if (stage === null) {
-        return (
-            <NewStageForm
-                onSubmitted={onSubmitted}
-                onCancel={onCancel}
-                existingStages={existingStages}
-            />
-        )
+        return <NewStageForm onSubmitted={onSubmitted} onCancel={onCancel} />
     }
 
     return (
@@ -422,7 +401,6 @@ export const EditOrNewStageForm = ({
             stage={stage}
             onCancel={onCancel}
             onSubmitted={onSubmitted}
-            existingStages={existingStages}
         />
     )
 }