+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# [0.112.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.111.0...v0.112.0) (2026-01-21)
+
+
+### Features
+
+* predictors form [DHIS2-20616] ([#725](https://github.com/dhis2/maintenance-app-beta/issues/725)) ([8cec70e](https://github.com/dhis2/maintenance-app-beta/commit/8cec70e3460563e770d7b1d69cf57a79c16cffb4))
+
+# [0.111.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.110.0...v0.111.0) (2026-01-20)
+
+
+### Features
+
+* [DHIS2-20632] Program tracked entity attributes ([#720](https://github.com/dhis2/maintenance-app-beta/issues/720)) ([a492eca](https://github.com/dhis2/maintenance-app-beta/commit/a492eca7bb19ecb3ed26d6c00f7fc4d94295acd9))
+
+# [0.110.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.109.1...v0.110.0) (2026-01-14)
+
+
+### Features
+
+* [DHIS2-20479] Program access and sharing ([#716](https://github.com/dhis2/maintenance-app-beta/issues/716)) ([994cf5a](https://github.com/dhis2/maintenance-app-beta/commit/994cf5a106b807ff18eb3805f0500b8fe3ebd3fc))
+
+## [0.109.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.109.0...v0.109.1) (2026-01-14)
+
+
+### Bug Fixes
+
+* add validation strategy field to program stage form and fix bug ([#730](https://github.com/dhis2/maintenance-app-beta/issues/730)) ([f41505a](https://github.com/dhis2/maintenance-app-beta/commit/f41505a3acc2c9504472141b7c09d6555d830360))
+
+# [0.109.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.108.1...v0.109.0) (2026-01-14)
+
+
+### Features
+
+* [DHIS2-20618] Program rule variable list ([#728](https://github.com/dhis2/maintenance-app-beta/issues/728)) ([8d23fb7](https://github.com/dhis2/maintenance-app-beta/commit/8d23fb7323c6677698d731324a79252e03957a24))
+
 ## [0.108.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.108.0...v0.108.1) (2026-01-12)